Revert "created polling on lobby page, create submit definition page, modifie…"
