+++ balderdash-app/public/components/horizontal-container-v.js
@@ -39,14 +39,15 @@ export class HorizontalContainerV extends BaseContainer {
     const maxWidth = this.getAttribute('maxWidth') || '';
     const minHeight = this.getAttribute('minHeight') || '';
     const alpha = this.getAttribute('alpha') || 1;
-    const justifyContent = this.getAttribute('justifyContent') || 'center';
-    
+
+
+
     return `
     .container {
       flex-direction: row;
       flex: 1 0 auto;
+      justify-content: center;
       align-items: center;
-      ${justifyContent ? `justify-content: ${justifyContent};` : ''}
       ${backgroundColour ? `background-color: ${hexToRgba(backgroundColour, alpha)};` : ''}
       ${margin ? `margin: ${margin};` : ''}
       ${padding ? `padding: ${padding};` : ''}
