+++ balderdash-app/public/js/apiClient.js
@@ -1,10 +1,8 @@
 import { showErrorScreen } from "./helpers.js";
-import { clearStorage } from "./storage.js";
 
 const API_BASE_URL = window.location.origin + '/api'; 
 
 export async function logoutUser() {
-  clearStorage();
   await fetch(window.location.origin+'/logout');
   window.location.href = '/sign-in';
 }
